+++ Changelog.md
@@ -1,3 +1,6 @@
+## 0.0.55
++ Fixed `fxn.predictions.create` method failing due to downloaded resources being used by more than one process.
+
 ## 0.0.54
 + Added support for targeting `htp` inference backend in `beta.QnnInferenceMetadata`.
 + Added `index_url` argument in `Sandbox.pip_install(...)` method to specify Python package index URL.